Extract logic that starts audio recording Activity
