+++ collect_app/src/main/java/org/odk/collect/android/audio/AudioHelper.java
@@ -10,6 +10,7 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.OnLifecycleEvent;
 import androidx.lifecycle.ViewModelProviders;
 
+import org.odk.collect.android.widgets.utilities.AudioPlayer;
 import org.odk.collect.async.Scheduler;
 import org.odk.collect.audioclips.AudioClipViewModel;
 import org.odk.collect.audioclips.Clip;
