+++ collect_app/src/main/java/org/odk/collect/android/audio/ViewModelAudioPlayer.java
@@ -3,6 +3,7 @@ package org.odk.collect.android.audio;
 import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.Observer;
 
+import org.odk.collect.android.widgets.utilities.AudioPlayer;
 import org.odk.collect.audioclips.AudioClipViewModel;
 import org.odk.collect.audioclips.Clip;
